refactor controller product, cashback and order
