+++ src/adapters/cashback/cashback.ts
@@ -1,0 +1,8 @@
+import {
+  createCashback as createCashbackCore,
+  CreateCashback,
+} from '@/core/use-cases/cashback/create-cashback'
+
+export const createCashback: CreateCashback = (saveCashback) => (cashback) => {
+  return createCashbackCore(saveCashback)(cashback)
+}
