+++ src/adapters/index.ts
@@ -1,0 +1,3 @@
+export * from './cashback/cashback'
+export * from './product/product'
+export * from './order/order'
