+++ src/core/use-cases/order/create-order.ts
@@ -1,26 +1,13 @@
 import * as TE from 'fp-ts/TaskEither'
 import { pipe } from 'fp-ts/function'
-import { right, Either, left, toError } from 'fp-ts/Either'
+import { toError } from 'fp-ts/Either'
 import { Order } from '@/core/types/order'
+import { validateOrder } from './validateOrder'
 
 export type SaveOrder = (o: Order) => Promise<unknown>
-type CreateOrder = (f: SaveOrder) => (o: Order) => TE.TaskEither<Error, unknown>
+export type CreateOrder = (f: SaveOrder) => (o: Order) => TE.TaskEither<Error, unknown>
 
-const hasProduct = (order: Order): boolean => {
-  return order.productList.length > 0
-}
-const isProductsValid = (order: Order): boolean => {
-  const orderValidators = [hasProduct(order)]
-  return orderValidators.every((item) => item === true)
-}
-const validateOrder = (order: Order): Either<Error, Order> => {
-  if (isProductsValid(order)) {
-    return right(order)
-  }
-  return left(new Error('Invalid Product! - No products in the product list.'))
-}
-
-export const createOrder: CreateOrder = (saveOrder) => (order: Order) => {
+export const createOrder: CreateOrder = (saveOrder) => (order) => {
   return pipe(
     order,
     validateOrder,
@@ -31,3 +18,12 @@ export const createOrder: CreateOrder = (saveOrder) => (order: Order) => {
     )),
   )
 }
+
+/*
+const keys = Object.keys(req.body)
+for (const key of keys) {
+  if (req.body[key] === '') {
+    return res.status(400).json(createErrorMessage('Please, fill all fields!'))
+  }
+}
+*/
