+++ src/core/use-cases/product/create-product.ts
@@ -1,10 +1,10 @@
+import { toError } from 'fp-ts/Either'
 import * as TE from 'fp-ts/TaskEither'
 import { pipe } from 'fp-ts/function'
-import { toError } from 'fp-ts/Either'
 import { Product } from '@/core/types/product'
 
 export type SaveProduct = (p: Product) => Promise<unknown>
-type CreateProduct = (f: SaveProduct) => (p: Product) => TE.TaskEither<Error, unknown>
+export type CreateProduct = (f: SaveProduct) => (p: Product) => TE.TaskEither<Error, unknown>
 
 export const createProduct: CreateProduct = (saveProduct) => (product) => {
   return pipe(
@@ -14,3 +14,17 @@ export const createProduct: CreateProduct = (saveProduct) => (product) => {
     ),
   )
 }
+
+/*
+const keys = Object.keys(req.body)
+
+for (const key of keys) {
+  if (req.body[key] === '') {
+    return res.status(400).json(createErrorMessage('Please, correctly fill fields!'))
+  }
+}
+
+if (!isNumber(req.body.price)) {
+  return res.status(400).json(createErrorMessage('Please, correctly fill fields!'))
+}
+*/
